requirements.txt
asyncio==3.4.3
colorama==0.4.3
discord.py==1.4.1
nekos.py==1.0.3
nest-asyncio==1.4.0
wikipedia==1.4.0
mcstatus==4.0.0
beautifulsoup4==4.9.1
youtube-dl==2020.12.12
google-api-python-client==1.12.3
PyNaCl==1.4.0
pymongo==3.11.0
dnspython==1.16
requests==2.24.0
numba==0.52.0